added a bit to the fishing code
The fishing code is garbage and will likely need to be completely reworked or a creative fix needs to be found. Set sleep delays can provide a semi-reliable cast (Casting 48-52 MOST of the time) but even checking for a certain RGB color or the '100%' cue on fish meter screen to initiate a catch during fishing is unreliable at best.