Licensed Bit-HeroesBot
The software and files in this repository or "project" are subject to the GNU GPL V3 LICENSE. The license is an agreement. Violating this agreement can constitute breach of contract. I believe in the user's rights and free software. If you have any intent to fork, copy, rework or in any way use or change the code or software here in this repository, especially for commercial use, please make sure you have a clear understanding of this license.

Everything you need to know and any clarification about your rights and restrictions (there aren't that many) under the GNU GPL v3 license can be found at the links provided
https://www.gnu.org/licenses/gpl-3.0.en.html
https://www.fsf.org/